requirements.txt
# requirements.txt
paho-mqtt
Pillow
PyYAML
aioesphomeapi
